Add matmul for fp16 using cublas

+++ cuda_mode/matmul/src/run_fp16.cuh
@@ -13,4 +13,4 @@ void check_cuda(cudaError_t err, const char* const func, const char* const file,
 
 void cuda_device_info();
 
-void run_kernel_fp16(int kernel_num, int m, int n, int k, float alpha, float *A, float *B, float beta, float *C, cublasHandle_t handle);
+void run_kernel_fp16(int kernel_num, int m, int n, int k, half alpha, half *A, half *B, half beta, half *C, cublasHandle_t handle);
